Calm with a login system and more express pull on the backend
This commit moves all of the user handling into express and that's just
dandy. It handles sessions and it handles required authentication. The
login model will like Instapaper, where you only need an email to get
started, and then you can add a password for added security. Here the
users.authenticate method has been modified to take password optionally,
though I would like for everything to just work off an IP or a session
w/o a user attached to it. We'll see..